Camera now influences indicator with new DetectionManager
A new Object prefab and script has been created called DetectionManager. this object functions as a singleton and manages detection between multiple types of enemies. It contains a list of all objects currently detecting the player and outputs an event everytime an object gets added to this list and one when the list is emptied.
This object is REQUIRED for the security camera to function. When this object doesn't exist in a scene with a security camera the game will throw an error popup.
Furthermore the DetectionIndicator script has been modified to react to the DetectionManager Events.
Final minor changes include:
- Global Volume is now a prefab that can be easily dragged into new scenes
- The Security Camera has been made into a prefab.